Supported namespace (ns)

+++ Makefile
@@ -3,17 +3,21 @@ CONSTRUCT=tools/jsconstruct
 COMPILE_ARC=bin/arcjsc
 DATETIME=$(shell date +'%Y%m%d%H%M%S')
 BUILDDIR=.
+CURRENT_BRANCH=$(shell git branch | grep \* | sed -e "s/^\* //")
 
 all:		arc.js arc.min.js
 
 clean:
 		rm -f arc.js arc.min.js
 
+unit:
+		mocha -b test/unit/unit.js
+
 auto:
 		make
 		./tools/update-watcher src/*.js src/arc/*.arc -- make
 
-arc.js:		src/*.js src/compiler.fasl src/arc.fasl
+arc.js:		src/*.js src/primitives/*.js src/compiler.fasl src/arc.fasl
 	        $(CONSTRUCT) -o arc.tmp.js src/arc.js
 		cat src/comments.js arc.tmp.js > arc.js
 		rm arc.tmp.js
@@ -26,11 +30,23 @@ arc:		src/arc/arc.arc
 
 compiler:       src/arc/compiler.arc
 		mkdir -p backup
-		cp src/compiler.fasl backup/$(DATETIME).compiler.fasl
+		if [ -e src/compiler.fasl ]; then cp src/compiler.fasl backup/$(DATETIME).compiler.fasl; fi;
 		$(COMPILE_ARC) -d -o src/compiler.fasl src/arc/compiler.arc
 
 restore_compiler:
 		mv $(shell ls backup/*.compiler.fasl | sort -r | sed '1!d') src/compiler.fasl
 		make
+
+install_web:	arc.min.js
+		cp arc.js arc.js.bk
+		cp arc.min.js arc.min.js.bk
+		git checkout gh-pages
+		mv arc.js.bk js/arc.js
+		mv arc.min.js.bk js/arc.min.js
+		git add js/
+		git commit -m 'Installed new arc.js.'
+		git checkout $(CURRENT_BRANCH)
+		rm -f arc.js.bk arc.min.js.bk
+
 
 .PHONY:		all clean
